Minor changes with respect to previous commit Added a few lines of comments, and organized the code a little bit. No major changes. The only real change was the step where we "de-integrate" the spectral function (couldn't find a better way to say that). That is, in the SF file, the probability values "S" are given already integrated as follows:
S = 4*Pi [s] Pm^2 dPm dEm

where s is the actual ("un-integrated") probability. In other words, we are given S and we actually need s. Thus, we have to do:

s = S / (4*Pi Pm^2 dPm dEm)

This step was done in a step that was not very convenient. We moved it and now it is done as soon as the normalization is calculated (Note: this MUST be done after the normalization is computed, otherwise the results are simply wrong!)