Added a new branch to the output tree I was looking at the in-plane angle of the outgoing proton, and I was doing the conversion from h_yptar (ssyptar) to "theta" by hand. I decided to add a new branch that outputs theta in degrees directly. To do this, I had to implement two small changes:
1) In: NtupleInit.f I added two lines:
	From:
	m = m+1
        NtupleTag(m) = 'SF_weight_recon'      ! 66 spectral function with reconstructed quantities

	To:
	m = m+1
        NtupleTag(m) = 'SF_weight_recon'      ! 66 spectral function with reconstructed quantities
        m = m+1                     ! 67 RCT 8/9/2016 outgoing proton in-plane angle
        NtupleTag(m) = 'h_Thf'      ! 67 RCT 8/9/2016 outgoing proton in-plane angle

2) In: results_write.f I added one line
	From:
        ntu(67) = (spec%p%theta+recon%p%yptar)*180./3.1415926536     !  RCT 8/9/2016 outgoing reconstructed proton in-plane angle

	To:
        ntu(66) = main%SF_weight                     !  Spectral Function fro recon. quantities
        ntu(67) = (spec%p%theta+recon%p%yptar)*180./3.1415926536     !  RCT 8/9/2016 outgoing reconstructed proton in-plane angle

I did this following Werner's email:
"Here is (in short) how to add a variable to the output tree. This is for e,e’p only:
In NtupleInit.f add another NtupleTag (the last once currently is SF_weight_recon)
and make sure you increment the counter variable m.
Then in results_write.f you store the corresponding value in the ntu array. The way
it is programmed is a bit dangerous as here you need to give the actual array index
.e.g. 67, in other words you need to increment it yourself adn hardwire it. This could
be changed but I was too lazy to change the original code.
This is basically all you have to do. The branches in the tree are created automatically,
so nothing else needs to be changed."