Corrected issue with missing momentum Following Barak's email.
Hi Rey,

I looked a little at the missing momentum components, and there was something
strange going on. In the event.f code, the (x,y,z) components are in the lab
frame (with z downstream, x vertically down, y horizontal). But in the
results_write.f code, they seem to change the coordinate system before writing
out to the ntuple. These new variables are not used anywhere else; and they
are similar to the ones relative to the q vector. So, I'm not quite sure what
use they are. So, I adjusted it back to the lab frame. Here's the code:

       if (doing_hyd_elast .or. doing_deuterium .or. doing_heavy) then
          poftheta = Mp*Ebeam / (2*ebeam*sin(recon%e%theta/2.)**2 + Mp)
          corrsing = recon%e%P - poftheta
!         Pm_Heepz = -(recon%Pmy*recon%uq%y+recon%Pmz*recon%uq%z)
!     >         / sqrt(recon%uq%y**2+recon%uq%z**2)
!         Pm_Heepy =  (recon%Pmz*recon%uq%y-recon%Pmy*recon%uq%z)
!     >         / sqrt(recon%uq%y**2+recon%uq%z**2)
!         Pm_Heepx =  -recon%Pmx

          !Keep in lab coordinate system. Barak Schmookler, March 2017
          Pm_Heepz = recon%Pmz
          Pm_Heepy = recon%Pmy
          Pm_Heepx = recon%Pmx
        endif

Also note that in SIMC they calculate the missing momentum as p_proton - p_q,
which is negative the recoil momentum.
-Barak